Move DAO related classes/interfaces from the middle package to their own dao package.
