+++ customer/src/main/java/CustomerModel.java
@@ -3,7 +3,7 @@ import products.exceptions.ProductDoesNotExistException;
 import javafx.beans.property.SimpleStringProperty;
 import products.ProductDAO;
 import middle.MiddleFactory;
-import middle.DAOException;
+import dao.DAOException;
 import products.usecases.GetProductByNumber;
 import products.usecases.GetProductsBySearch;
