+++ shared/src/main/java/dao/DAOException.java
@@ -1,4 +1,4 @@
-package middle;
+package dao;
 
 import java.io.Serial;
