+++ shared/src/main/java/dao/RemoteDAOWrapper.java
@@ -1,4 +1,4 @@
-package middle;
+package dao;
 
 import java.rmi.RemoteException;
 import java.util.List;
