+++ shared/src/main/java/dao/RemoteMethod.java
@@ -1,4 +1,4 @@
-package middle;
+package dao;
 
 import java.rmi.RemoteException;
