+++ shared/src/main/java/products/remote/RemoteProductDAO.java
@@ -1,6 +1,6 @@
 package products.remote;
 
-import middle.RemoteDAO;
+import dao.RemoteDAO;
 import products.Product;
 
 public interface RemoteProductDAO extends RemoteDAO<Product> {}
