+++ shared/src/main/java/products/usecases/RestockProduct.java
@@ -3,7 +3,7 @@ package products.usecases;
 import products.Product;
 import products.exceptions.ProductDoesNotExistException;
 import products.ProductDAO;
-import middle.DAOException;
+import dao.DAOException;
 
 public class RestockProduct {
     private final ProductDAO productDAO;
